+++ src/main/java/org/sonar/plugins/delphi/pmd/rules/MethodResultAssignedRule.java
@@ -1,29 +1,27 @@
 package org.sonar.plugins.delphi.pmd.rules;
 
+import static org.sonar.plugins.delphi.utils.MethodUtils.isMethodStubWithStackUnwinding;
+import static org.sonar.plugins.delphi.utils.StatementUtils.isMethodInvocation;
+
 import java.util.ArrayDeque;
-import java.util.Collections;
 import java.util.Deque;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import net.sourceforge.pmd.RuleContext;
-import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import org.sonar.plugins.delphi.antlr.ast.node.ArgumentListNode;
 import org.sonar.plugins.delphi.antlr.ast.node.AsmStatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.AssignmentStatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.DelphiNode;
 import org.sonar.plugins.delphi.antlr.ast.node.ExpressionNode;
-import org.sonar.plugins.delphi.antlr.ast.node.ExpressionStatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.ForLoopVarReferenceNode;
 import org.sonar.plugins.delphi.antlr.ast.node.FormalParameterNode.FormalParameterData;
 import org.sonar.plugins.delphi.antlr.ast.node.IdentifierNode;
 import org.sonar.plugins.delphi.antlr.ast.node.MethodBodyNode;
 import org.sonar.plugins.delphi.antlr.ast.node.MethodImplementationNode;
 import org.sonar.plugins.delphi.antlr.ast.node.NameReferenceNode;
-import org.sonar.plugins.delphi.antlr.ast.node.PrimaryExpressionNode;
-import org.sonar.plugins.delphi.antlr.ast.node.RaiseStatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.StatementNode;
 import org.sonar.plugins.delphi.antlr.ast.node.UnaryExpressionNode;
 import org.sonar.plugins.delphi.operator.UnaryOperator;
@@ -110,62 +108,8 @@ public class MethodResultAssignedRule extends AbstractDelphiRule {
     return body == null || isMethodStubWithStackUnwinding(method) || body.hasAsmBlock();
   }
 
-  private static boolean isMethodStubWithStackUnwinding(MethodImplementationNode method) {
-    MethodBodyNode body = method.getMethodBody();
-    if (body.hasStatementBlock()) {
-      for (StatementNode statement : body.getStatementBlock().getStatements()) {
-        if (isStackUnwindingStatement(statement)) {
-          return true;
-        } else if (!(statement instanceof AssignmentStatementNode)) {
-          return false;
-        }
-      }
-    }
-    return false;
-  }
-
-  private static boolean isStackUnwindingStatement(StatementNode statement) {
-    return statement instanceof RaiseStatementNode || isAssertFalse(statement);
-  }
-
   private static boolean isExitWithReturnValue(StatementNode statement) {
     return isMethodInvocation(statement, "System.Exit", arguments -> arguments.size() == 1);
-  }
-
-  private static boolean isAssertFalse(StatementNode statement) {
-    return isMethodInvocation(statement, "System.Assert", arguments -> arguments.get(0).isFalse());
-  }
-
-  private static boolean isMethodInvocation(
-      StatementNode statement,
-      String fullyQualifiedName,
-      Predicate<List<ExpressionNode>> argumentListPredicate) {
-    if (!(statement instanceof ExpressionStatementNode)) {
-      return false;
-    }
-
-    var expression = ((ExpressionStatementNode) statement).getExpression().skipParentheses();
-    if (!(expression instanceof PrimaryExpressionNode) || expression.jjtGetNumChildren() > 2) {
-      return false;
-    }
-
-    Node name = expression.jjtGetChild(0);
-    if (!(name instanceof NameReferenceNode)) {
-      return false;
-    }
-
-    NameDeclaration declaration = ((NameReferenceNode) name).getLastName().getNameDeclaration();
-    List<ExpressionNode> arguments = extractArguments(expression.jjtGetChild(1));
-    return declaration instanceof MethodNameDeclaration
-        && ((MethodNameDeclaration) declaration).fullyQualifiedName().equals(fullyQualifiedName)
-        && argumentListPredicate.test(arguments);
-  }
-
-  private static List<ExpressionNode> extractArguments(Node argumentList) {
-    if (argumentList instanceof ArgumentListNode) {
-      return ((ArgumentListNode) argumentList).getArguments();
-    }
-    return Collections.emptyList();
   }
 
   private boolean handlePascalResultAssignments(NameReferenceNode nameReference) {