+++ src/main/java/org/sonar/plugins/delphi/symbol/scope/MethodScope.java
@@ -1,8 +1,10 @@
 package org.sonar.plugins.delphi.symbol.scope;
 
 import javax.annotation.Nullable;
+import org.sonar.plugins.delphi.symbol.declaration.MethodNameDeclaration;
 
 public class MethodScope extends AbstractDelphiScope {
+  private MethodNameDeclaration methodNameDeclaration;
   private DelphiScope typeScope;
 
   @Nullable
@@ -12,6 +14,15 @@ public class MethodScope extends AbstractDelphiScope {
 
   public void setTypeScope(DelphiScope typeScope) {
     this.typeScope = typeScope;
+  }
+
+  @Nullable
+  public MethodNameDeclaration getMethodNameDeclaration() {
+    return methodNameDeclaration;
+  }
+
+  public void setMethodNameDeclaration(MethodNameDeclaration methodNameDeclaration) {
+    this.methodNameDeclaration = methodNameDeclaration;
   }
 
   @Override